+++ packages/in_app_purchase/in_app_purchase_ios/lib/src/types/types.dart
@@ -1,0 +1,7 @@
+// Copyright 2013 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+//
+export 'app_store_product_details.dart';
+export 'app_store_purchase_details.dart';
+export 'app_store_purchase_param.dart';