+++ packages/google_maps_flutter/google_maps_flutter/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.0.5
+
+* Google Maps requires at least Android SDK 20.
+
+## 2.0.4
+
+* Unpin iOS GoogleMaps pod dependency version.
+
 ## 2.0.3
 
 * Fix incorrect typecast in TileOverlay example.